Fixed the page shift problem, updated utils with UUID generating function and a ISOLocal time function and discovered and dealt with a new date problem
