+++ src/lib/utils.ts
@@ -59,3 +59,25 @@ export const flyAndScale = (
     easing: cubicOut
   };
 };
+
+
+export function toISOLocal(date: string): string {
+  const parts = date.match(/(\d+)/g);
+  if (!parts || parts.length < 5) return '';
+  const adate = new Date(
+    Number(parts[2]),
+    Number(parts[1]) -1,
+    Number(parts[0]),
+    Number(parts[3]),
+    Number(parts[4])
+  );
+  const localdt = new Date(adate.getTime() - adate.getTimezoneOffset() * 60000);
+  return localdt.toISOString().slice(0, -1);
+}
+
+export function guidGenerator() {
+  var S4 = function() {
+    return (((1+Math.random())*0x10000)|0).toString(16).substring(1);
+  };
+  return (S4()+S4()+"-"+S4()+"-"+S4()+"-"+S4()+"-"+S4()+S4()+S4());
+}
