+++ src/routes/(app)/team/edit/[id]/+page.server.ts
@@ -18,7 +18,7 @@ export const actions = {
     const name = data.get('name');
     const bio = data.get('bio');
     const birthDate = new Date(data.get('birthDate') as string);
-    const birthDatejson = `${birthDate.getDay()}-${birthDate.getMonth()}-${birthDate.getFullYear()} ${birthDate.getHours()}:${birthDate.getMinutes()}`;
+    const birthDatejson = `${birthDate.getDate()}-${birthDate.getMonth()+1}-${birthDate.getFullYear()} ${birthDate.getHours()}:${birthDate.getMinutes()}`;
     const linkedin = data.get('linkedin');
     const github = data.get('github');
     const photo = data.get('photo') as File;
